backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
django-environ>=0.11
django-cors-headers>=4.3
django-filter>=24.1
gunicorn>=22
whitenoise>=6.7
Pillow>=10
django-unfold>=0.40
psycopg[binary]>=3.2
